Bypass seconadry storage
To bypass secondary storage the configuration settings "sp.bypass.secondary.storage" has to be set to "true".
We are handling the creation of template from snapshot in StorPoolDataMotionStrategy.canHandle command and in StorPoolDataMotionStrategy.copyAsync a StorPool's volume is created from the snapshot and freezed (snapshoted). That's how we keep it in StorPool than secondary storage.
To bypass secondary storage for backup snapshots we lie CloudStack that the snapshot is backuped to secondary and adding a row in snapshot_store_ref table.

Change-Id: Ib871db02559eca142d2fffccaafb8ab809ecce39