Backup template on secondary storage created from StorPool snapshot
Allow download template to work when secondary storage bypass is enabled. Assumes that the number of created templates will be small overall and that this would not consume significant space on the secondary storage. When creating template from snapshot and "sp.bypass.secondary.storage" is set to true, we convert the template to QCOW and save it to secondary. There will be snapshot for the template on StorPool which will be set in DB table cloud.template_store_ref on install_path. Catch of DeleteTemplateCmd to delete our template snapshot, when template is deleted on secondary

Change-Id: I7a2306e6f020c6858995da9abaa0f3ec452a18c7