mgmt: Fix in create VM if snapshot is not on StorPool
CloudStack is cleaning templates on primary, if there are no active
volumes using them, but there is a potential race - the snapshot can get
deleted in storpool but the record in tempate_spool_ref to be still
present - e.g. restart of cloudstack mgmt after the StorPool API call
has been issued, but before the database was actually updated.To handle
this and other similar cases - e.g. someone deleting the snapshot by hand
in StorPool by mistake, we'll recreate it again on StorPool by cleaning
first the template_spool_ref table.

Change-Id: I594938f6e9f6f2c0272bea426733af982fa213b6