mgmt/agent: Change StorPool's snapshots/volumes names with globalId
Migration from UUID names to globalIds. Names for all objects
snapshots/volumes will be created from StorPool as a globalIds.
The reason for the change is when the cloudstack controller is restarted while
template is copied from secondary to storpool primary. The end result
is that there are both a volume and a snapshot with the same name in the
multicluster, and that isn't supposed to happen (but is not enforced by
storpool, as it will block local volume+snapshot operations in a cluster
if another cluster is down). The solution is to delete the volume (or just
rename it to be sure) as it is from the old not completed attempt to copy
the snapshot to StorPool. With this change in the plugin, it will not "name"
the volumes and snapshots but use only the globally unique ids that StorPool
generates internally for each volume and snapshot. (The names will be stored
in tags for reference, but will not be used by the plugin). This change also
fixes the issue at hand as there can't be a collision with the new approach.

Change-Id: I89ff9211c37fa290666fbc0a2cfc896c1381e009