mgmt: handle left objects if deleting primary storage
When deleting StorPool's primary storage we handle if there are
active snapshots and snapshots of a templates. This will prevent:
1 - to leave active snapshots and templates on CloudStack and StorPool.
And when is the primary storage is deleted, you cannot delete them from
Cloudstack, it will throw an exception that there is no such primary
storage.
2 - The snapshots will be left on StorPool, and if you want to delete
StorPool's template, you have to clean it manualy from those snapshots.
3- After the handling if there are snapshots and templates Cloudstack
won't let you delete the primary storage

Change-Id: I6d6c5cc91190aae664d5518381a1874f439c96d7