mgmt/agent: support all CloudStack released versions
Removed override annotation to methods which appear in newew versions
and on methods which arguments are changed (will use the appropriate one
depends on version). Replaced functionality which is deprecated in older
versions of CS and is removed in newer ones. Implementing
SnapshotStrategy, because in newer versions the class that we had
extending is renamed

Change-Id: I3adb8fe269e1f4c9c075147d155846c7831c03e1