agent: workaround for storage capacity
There is a fix in CS to update storage capacity when storage stats is
invoked (in our case was set to 0).
Returning the capacity of the template won't work, because it won't take
into account the thin provisioning, and there is no good way to handle
multi-cluster deployments (where there is more than one actual storage).

Change-Id: I85ab7783dc6f4e3c772c9f62800c94a89e38375f