mgmt: Fix volume creation from snapshot with secondary storage
If `sp.bypass.secondary.storage` is set we lie CS that the
snapshot exists on secondary storage. When we create volume from
snapshot if for some reason the snapshot is deleted from StorPool we'll
try to create volume from secondary storage snapshot, and we'll check if
it's realy exists on it. Otherwise the command will fail

Change-Id: Ia25ece6e2e490b2d59a04b57f40d882b08e0c44e