mgmt: Fix attach volume
When a VM is deployed with the option `startvm=false` the volumes are in
Allocated state. This means that are not created on primary storage. When you
try to attach a newly created volume in Allocated state to such a VM,
the volume is attached to it but it never gets created on the storage. This is
the default CS behavior no matter what the storage is. In each case we are
trying to get the path of the volume but in this it's null and a
NullPointerException is thrown

Change-Id: I78a6f0f716bc4130e0bb5e9555f87021620da101