mgmt/agent/marvin: Support partial Zone-Wide storage
If there is no StorPool installed on the agent, or it's not setup correctly,
we'll remove any record for this pool from the `storage_pool_host_ref` table.
This will enable partial Zone-wide storage pool until StorPool
is setup correctly on the host or if it is never installed.

Also refactor Marvin tests.

Change-Id: I09769c23ebc8ccce4ecf9fa98badeab99c950443