agent: Support for direct download of a template
Support to download a QCOW2 template by skipping the usage of secondary
storage. After a template is register with the option `directdownload`
and used for first time to deploy a virtual machine, CloudStack will
choose on random a host to download the template and will be converted
to a StorPool snapshot for further use.

Change-Id: I08b4c2aa2f3090a51f4b5ab6970f8f638b9b0c0b